+++ android/src/main/java/com/razorpay/flutter_customui/RazorpayPlugin.java
@@ -61,10 +61,14 @@ public class RazorpayPlugin  implements FlutterPlugin, MethodCallHandler, Activi
 
   @Override
   public void onMethodCall(@NonNull MethodCall call, @NonNull final Result result) {
+    if (razorpayDelegate == null) {
+      result.error("UNAVAILABLE", "RazorpayDelegate not initialized", null);
+      return;
+    }
 
     switch (call.method) {
       case "initilizeSDK":
-        razorpayDelegate.init(call.arguments.toString(),result);
+        razorpayDelegate.init(call.arguments.toString(), result);
         break;
 
       case "submit":
@@ -117,6 +121,10 @@ public class RazorpayPlugin  implements FlutterPlugin, MethodCallHandler, Activi
 
       case "setPaymentId":
         razorpayDelegate.setPaymentID(call.arguments.toString(), result);
+        break;
+      
+      case "resync":
+        razorpayDelegate.resync(result);
         break;
 
       //Turbo UPI
@@ -193,22 +201,14 @@ public class RazorpayPlugin  implements FlutterPlugin, MethodCallHandler, Activi
         break;
 
       default:
-        Log.d(TAG,"no method");
+        result.notImplemented();
     }
   }
 
   @Override
   public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
     channel.setMethodCallHandler(null);
-    eventChannel.setStreamHandler(null);
-    this.eventChannel.setStreamHandler(null);
-  }
-
-  @RequiresApi(api = Build.VERSION_CODES.KITKAT)
-  public RazorpayPlugin(Registrar registrar) {
-    this.activity = registrar.activity();
-    this.razorpayDelegate = new RazorpayDelegate(registrar.activity());
-    registrar.addActivityResultListener(razorpayDelegate);
+    channel = null;
   }
 
   @RequiresApi(api = Build.VERSION_CODES.KITKAT)
@@ -242,8 +242,11 @@ public class RazorpayPlugin  implements FlutterPlugin, MethodCallHandler, Activi
 
   @Override
   public void onDetachedFromActivity() {
-    pluginBinding.removeActivityResultListener(razorpayDelegate);
-    pluginBinding = null;
+    if (pluginBinding != null) {
+      pluginBinding.removeActivityResultListener(razorpayDelegate);
+      pluginBinding = null;
+    }
+    razorpayDelegate = null;
   }
 
   // Turbo UPI