+++ CHANGELOG.md
@@ -50,3 +50,10 @@
 
 - Removed JCenter from build.gradle for android due to Deprecation
 
+## 1.4.0
+
+- Migrated to android v2 embedding.
+
+## 1.4.1
+
+- Bug-fix : Fixed `reply already submitted` intermittent error on getPaymentMethods function.